fixed bug and ui variables
there was a bug which was making the moore system just not work at all for some reason. had to do with the new neighbour checking system.

i also made some changes to the code to work with a few variables which will be used in the ui -
color gradient
 - two colors for gradient
 - color mode
  - dist from center or stage
 start state
 - either noise or a box in the center
noise hasnt been implemented yet
 - the center size variable dictates size of the box in the center